Add a custom exception message

+++ src/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { IdeaDTO } from './idea.dto';
@@ -22,16 +22,22 @@ export class IdeaService {
     }
 
     async read(id: string) {
-        return await this.ideaRepository.findOne({ where: { id } })
+        const idea = await this.ideaRepository.findOne({ where: { id } })
+        if (!idea) throw new HttpException('Not found', HttpStatus.NOT_FOUND)
+        return idea
     }
 
     async update(id: string, data: Partial<IdeaDTO>) {
+        const idea = this.ideaRepository.findOne({ id })
+        if (!idea) throw new HttpException('Not found', HttpStatus.NOT_FOUND)
         await this.ideaRepository.update({ id }, data)
-        return await this.ideaRepository.findOne({ id })
+        return idea
     }
 
     async destroy(id: string) {
+        const idea = this.ideaRepository.findOne({ id })
+        if (!idea) throw new HttpException('Not found', HttpStatus.NOT_FOUND)
         await this.ideaRepository.delete({ id })
-        return { deleted: true }
+        return idea
     }
 }
